Fix issue 25: potential error when parsing big XML file
The split of the XML file may occur in the middle of a node name, and the context is not able to handle it. Now, we prevent such split. The parser will always keep some spare data for such cases by splitting earlier and unparsed data will be processed on next iteration.