Fix issue 52: roxml_del_node + roxml_add_node may hide and leak existing children
The last child of a parent node is always the last sibling. Thus,
when the last sibling is deleted, the last child of its parent
cannot be the next sibling (since we are already the last one it
means the next is always NULL) but the previous one.

Yet we cannot use roxml_get_prev_sibling() because this function
returns the previous XML element. We need to run through the
parent's children list in order to find the right sibling. This
is the purpose of the static function roxml_get_real_prev_sibling().

Signed-off-by: Emmanuel Deloget <logout@free.fr>