roxml_file: Fix context dependency on current line
The parsing of a file is done line after line, and the curr_name member of the context
is a pointer on that line.
If the pointer is set on one line and read on another one, its value in not constistent.

The fix consist in copying the value rather than pointing to it.
All the other context members are already independant of the line being parsed.