commit: introduce roxml_commit_fd()
The function is similar to roxml_commit_file() but allows one to commit
XML data to an already opened file. It's entirely possible to use special
file descriptors to write to (including, for example, the console out),
or to use multiple calls to this function with the same fd - meaning that
new XML data will be appended by each call.