roxml_parser: Replace previous parse engine with map based one
previous parse engine was linked list based.
To improve performances, this new parser is relying on a map.
There is no need to really hash the chunk for the state machine since we work with bytes.
The parser now uses a table with 512 (256 entries + 256 for future potential collision) to store callbacks
depending on the next byte available in the parsing string.

Quick performance testing using perf shows about 15% improvement on document load

With linked list based parser:
19.43%  roxml  [.] roxml_parse_line
10.59%  roxml  [.] _func_load_default
8.80%  roxml  [.] _func_load_close_node
6.17%  roxml  [.] roxml_del_tree

With map based parser
11.36%  roxml  [.] _func_load_default
10.19%  roxml  [.] _func_load_close_node
7.07%  roxml  [.] roxml_del_tree
3.92%  roxml  [.] roxml_parse_line