roxml_core: Fix potential memory leak.
When declaring a default namespace in a node that is inside a namespace that is not declared at all,
the default namespace was overriding the invalid one without freeing its temporary declaration.

ex:
<ns:node xmlsns='http://default'/>

In this case, 'ns' is not valid as not declared.

This is now detected and the temporary 'ns' definition is properly released.
A test is also added for this usecase.

Signed-off-by: Tristan Lelong <tristan@lelong.xyz>