roxml_core: fix silently dropped text at the end of xml document.
When a XML document ends with some text, this was silently dropped: not
included in the parsed tree without generating an error.
On top of that, the memory for the TXT node would still be allocated,
but never linked => memory leak.

This change now makes this use case a valid one (although the XML
document is not a valid document), but this:
- doesn't leak memory
- doesn't silently drop data

Also libroxml always accepted slightly invalid tress for convenience.

Signed-off-by: Tristan Lelong <tristan@lelong.xyz>